requirements.txt
flask~=3.1.0
requests~=2.32.3
gunicorn
schedule~=1.2.2

python-dotenv~=1.0.1
django-scheduler~=0.10.1
SQLAlchemy~=2.0.38